+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+- Restored Address field values which were not being saved (`formatted`, `raw`, `name`, `neighborhood`, `county`, and `placeId`). ([#158](https://github.com/doublesecretagency/craft-googlemaps/issues/158))
+
 ## 5.2.2 - 2026-05-11
 
 ### Fixed